requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-CORS==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
marshmallow==3.20.1
opencv-python-headless>=4.9.0
numpy>=1.26.0
